Move all the other libs too
