+++ lib/effect/SwingingLights.cpp
@@ -1,8 +1,7 @@
 #include "SwingingLights.hpp"
 
 #include <ColorPalette.hpp>
-
-#include "../../software/generic/Math.hpp"
+#include <Math.hpp>
 
 static const uint16_t SPREAD = MAX_UINT16 * 0.2;
