+++ software/generic/NetworkManager.hpp
@@ -1,9 +1,8 @@
 #ifndef __NETWORK_MANAGER_H__
 #define __NETWORK_MANAGER_H__
 
+#include <Radio.hpp>
 #include <Types.hpp>
-
-#include "Radio.hpp"
 
 class NetworkManager {
  public:
