+++ software/generic/test/FakeLedManager.hpp
@@ -1,11 +1,11 @@
 #ifndef __FAKE_LED_MANAGER_HPP__
 #define __FAKE_LED_MANAGER_HPP__
 
+#include <LedManager.hpp>
 #include <Types.hpp>
 #include <vector>
 
 #include "../../lib/effect/Effect.hpp"
-#include "../LedManager.hpp"
 
 class FakeLedManager : public LedManager {
  public:
