+++ software/generic/test/LedManagerTest.cpp
@@ -1,5 +1,6 @@
+#include <LedManager.hpp>
+
 #include "../../../lib/effect/Effects.hpp"
-#include "../LedManager.hpp"
 #include "DeviceDescription.hpp"
 #include "FakeLedManager.hpp"
 #include "FakeRadio.hpp"
