+++ software/generic/test/NetworkManagerTest.cpp
@@ -1,5 +1,6 @@
+#include <Radio.hpp>
+
 #include "../NetworkManager.hpp"
-#include "../Radio.hpp"
 #include "FakeRadio.hpp"
 #include "gtest/gtest.h"
