+++ software/generic/test/RadioStateIntegrationTest.cpp
@@ -1,5 +1,6 @@
+#include <Radio.hpp>
+
 #include "../NetworkManager.hpp"
-#include "../Radio.hpp"
 #include "../RadioStateMachine.hpp"
 #include "FakeNetwork.hpp"
 #include "FakeRadio.hpp"
