+++ software/generic/test/RadioStateMachineTest.cpp
@@ -1,5 +1,6 @@
+#include <Radio.hpp>
+
 #include "../NetworkManager.hpp"
-#include "../Radio.hpp"
 #include "../RadioStateMachine.hpp"
 #include "FakeLedManager.hpp"
 #include "FakeRadio.hpp"
